cam_application/requirements.txt
opencv_python==4.7.0.72
Pillow~=9.3.0
requests~=2.26.0
paho-mqtt~=1.6.1
cryptography~=35.0.0
pyttsx3~=2.90
python-dotenv~=0.19.1